+++ packages/rsi-vue-example/src/components/interpretation-player.vue
@@ -8,7 +8,6 @@
    <p> Widget: </p>
     <RsiApiInterpretationPlayer
       sdkKey="d9ed0b1e-6027-4c19-a43a-69b231713f1c" 
-      positionMenu="bottom" 
       :isPlayerControlled="true"
       v-on:onLanguageSelected="handleOnLanguageSelected"
       v-on:onReady="handleOnReady"