+++ packages/rsi-vue-example/src/router/index.js
@@ -1,13 +1,24 @@
 import Vue from 'vue'
-import VueRouter from 'vue-router';
-import Home from '@/views/home.vue';
+import Router from 'vue-router'
+import Home from '@/views/home.vue'
+Vue.use(Router)
 
-Vue.use(VueRouter);
-
-const routes = [
-  { path: '/', component: Home },
-];
-
-const router = new VueRouter({routes});
-
-export default router;
+export default new Router({
+  routes: [
+     {
+      path: '/',
+      name: 'home',
+      component: Home,
+    }, 
+    {
+      path: '/interpretation-manager-demo',
+      name: 'interpretation-manager-demo',
+      component: () => import('@/views/interpretation-manager-demo.vue')
+    },
+    {
+      path: '/interpretation-player-demo',
+      name: 'interpretation-player-demo',
+      component: () => import('@/views/interpretation-player-demo.vue')
+    },
+  ]
+})