+++ packages/rsi-vue-example/src/views/interpretation-manager-demo.vue
@@ -1,0 +1,19 @@
+<template>
+  <div>
+ <InterpretationManager/>  
+  </div>
+</template>
+
+<script>
+import InterpretationManager from '@/components/interpretation-manager.vue' 
+  export default {
+    components:{
+      InterpretationManager 
+    },
+    
+  }
+</script>
+
+<style lang="scss" scoped>
+
+</style>