+++ packages/rsi-base/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.16](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-base@0.2.14...@akkadu/rsi-base@0.2.16) (2021-07-02)
+
+**Note:** Version bump only for package @akkadu/rsi-base
+
+
+
+
+
+## [0.2.14](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-base@0.2.13...@akkadu/rsi-base@0.2.14) (2021-07-01)
+
+**Note:** Version bump only for package @akkadu/rsi-base
+
+
+
+
+
 ## 0.2.13 (2021-06-22)
 
 **Note:** Version bump only for package @akkadu/rsi-base