+++ packages/rsi-interpretation-player/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.21](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-player@0.3.18...@akkadu/rsi-interpretation-player@0.3.21) (2021-07-02)
+
+**Note:** Version bump only for package @akkadu/rsi-interpretation-player
+
+
+
+
+
+## [0.3.18](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-player@0.3.17...@akkadu/rsi-interpretation-player@0.3.18) (2021-07-01)
+
+**Note:** Version bump only for package @akkadu/rsi-interpretation-player
+
+
+
+
+
 ## 0.3.17 (2021-06-22)